Chatbot/requirements.txt
python-dotenv
faiss-cpu
google-generativeai
langchain
langgraph
langchain-google-genai